+++ test/data/simple.hpp
@@ -21,7 +21,7 @@ public:
   }
 
   mstch::node taxed_value() {
-    return static_cast<int>(m_value - (m_value * 0.4));
+    return m_value - (m_value * 0.4);
   }
 
   mstch::node in_ca() {
